Funcionalidad para eliminar un todo terminada

+++ src/todos/app.js
@@ -30,7 +30,6 @@ export const App = (elementId) => {
     // Referencias HTML
     const newDescriptionInput = document.querySelector(ElementIDs.NewTodoInput);
     const todoListUl = document.querySelector(ElementIDs.TodoList);
-    const todoDeleteBtn = document.querySelector(ElementIDs.todoDelete);
 
     // Listeners
     newDescriptionInput.addEventListener('keyup', (event) => {
@@ -49,11 +48,9 @@ export const App = (elementId) => {
     })
 
     todoListUl.addEventListener('click', (event) => {
-        const element = event.target.classList.contains('destroy')
-            ? event.target.closest('[data-id]') // Si el elemento clickeado tiene la clase 'destroy', busca el elemento más cercano con el atributo data-id
-            : null; // Si no, asigna null
-        if(!element) return;
-
+        const isDestroyElement = event.target.className === 'destroy'; // Verifica si el elemento clickeado tiene la clase 'destroy'
+        const element = event.target.closest('[data-id]'); // Busca el elemento más cercano que tenga el atributo data-id
+        if(!element || !isDestroyElement ) return;
         todoStore.deleteTodo(element.getAttribute('data-id')); // Obtiene el valor del atributo data-id y lo pasa a la función deleteTodo del store
         displayTodos();
     })
